로그인 정보 처리하는 훅 수정
